requirements.txt
aiohttp==3.8.1
deep_translator==1.8.3
langdetect==1.0.9
nextcord==2.1.0
nextcord_ext_menus==1.5.4
python-dotenv==0.20.0
requests==2.28.1
Telethon==1.24.0
